config/requirements.txt
# Alpaca 0DTE Trading Framework Requirements
# ==========================================

# Core Alpaca SDK and dependencies
alpaca-py>=0.29.0
requests>=2.30.0
pydantic>=2.0.3
pandas>=1.5.3
msgpack>=1.0.3
websockets>=10.4
sseclient-py>=1.7.2

# Data analysis and visualization
numpy>=1.21.0
scipy>=1.9.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Trading and backtesting
backtrader>=1.9.76
yfinance>=0.2.0

# Environment and configuration
python-dotenv>=1.0.0

# Development and testing
pytest>=7.1
pytest-asyncio>=0.23.7
requests-mock>=1.9.3
black>=24.3.0
isort>=5.10.1

# Optional: Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.0.0
